requirements.txt
bidict==0.22.1
click==8.1.3
Flask==2.2.3
Flask-SocketIO==4.3.2
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-dotenv==0.21.1
python-engineio==3.14.2
python-socketio==4.6.1
simple-websocket==0.9.0
six==1.16.0
SQLAlchemy==2.0.3
typing_extensions==4.5.0
Werkzeug==2.2.3
wsproto==1.2.0
